backend/requirements.txt
fastapi==0.111.1
starlette==0.37.2
uvicorn[standard]==0.23.2

pydantic==2.11.5
pydantic-core==2.33.2
python-dotenv==1.0.0
python-multipart>=0.0.9,<0.1
typing-extensions==4.14.0

openai==1.97.0
openai-agents==0.2.1
anyio>=4.6,<5
supabase==2.12.0

# Asset-Driven System Dependencies
psutil>=5.9.0                     # System resource monitoring for performance optimization
websockets>=12.0                  # WebSocket support for real-time updates
redis>=5.0.0                      # Optional caching layer for asset system
aiofiles>=23.0.0                  # Async file operations for asset processing
jinja2>=3.1.0                     # Template rendering for AI prompt generation
python-dateutil>=2.8.0            # Enhanced date/time handling for asset tracking
asyncpg>=0.29.0                   # PostgreSQL async driver (alternative to Supabase)
aioredis>=2.0.0                   # Async Redis client for caching
httpx>=0.25.0                     # HTTP client for external API calls
croniter>=1.4.0                   # Cron expression parsing for scheduled tasks
rich>=13.0.0                      # Enhanced console output for monitoring
prometheus-client>=0.19.0         # Metrics collection for monitoring
structlog>=23.0.0                 # Structured logging for better observability
pytest-asyncio>=0.23.0            # Async support for pytest
